refactor: move makeSchema into separate file from builder
